refactor: move CategoryDetail and CategoryMenu
- move components/category-detail and components/category-menu to nerdlets/mande-nerdlet/components
