+++ nerdlets/mande-nerdlet/MandeContainer.js
@@ -14,10 +14,10 @@ import {
   navigation,
 } from 'nr1'
 
-import CategoryMenu from '../../components/category-menu/CategoryMenu'
+import CategoryMenu from './components/category-menu/CategoryMenu'
 import MetricSidebar from '../../components/metric-sidebar/MetricSidebar'
 import MetricDashboard from '../../components/dashboard/MetricDashboard'
-import CategoryDetail from '../../components/category-detail/CategoryDetail'
+import CategoryDetail from './components/category-detail/CategoryDetail'
 import Selected from '../../components/metric-sidebar/Selected'
 import metricConfigs from '../../config/MetricConfig'
 import { FIND_USER_ATTRIBUTE } from '../../config/MetricConfig'
