+++ nerdlets/mande-nerdlet/components/category-detail/ChartGrid.js
@@ -14,9 +14,9 @@ import {
   PieChart,
   FunnelChart,
 } from 'nr1'
-import { dateFormatter } from '../../utils/date-formatter'
-import { formatFacets } from '../../utils/query-formatter'
-import { getHook } from '../../config/Hooks'
+import { dateFormatter } from '../../../../utils/date-formatter'
+import { formatFacets } from '../../../../utils/query-formatter'
+import { getHook } from '../../../../config/Hooks'
 
 const chartGrid = props => {
   const {
